functions/ocr_trigger/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    /home/vscode/.local/bin/pip-compile
#        functions/ocr_trigger/requirements.in
#        -o
#        functions/ocr_trigger/requirements.txt
#
annotated-types==0.7.0
    # via pydantic
anyio==4.12.1
    # via starlette
blinker==1.9.0
    # via flask
certifi==2026.1.4
    # via requests
cffi==2.0.0
    # via cryptography
charset-normalizer==3.4.4
    # via requests
click==8.3.1
    # via
    #   flask
    #   functions-framework
    #   uvicorn
cloudevents==1.12.0
    # via functions-framework
cryptography==46.0.5
    # via google-auth
deprecation==2.1.0
    # via cloudevents
exceptiongroup==1.3.1
    # via anyio
flask==3.1.3
    # via functions-framework
functions-framework==3.10.1
    # via -r functions/ocr_trigger/requirements.in
google-api-core==2.30.0
    # via google-cloud-documentai
google-auth==2.48.0
    # via
    #   google-api-core
    #   google-cloud-documentai
google-cloud-documentai==3.10.0
    # via -r functions/ocr_trigger/requirements.in
googleapis-common-protos==1.72.0
    # via
    #   google-api-core
    #   grpcio-status
grpcio==1.78.0
    # via
    #   google-api-core
    #   google-cloud-documentai
    #   grpcio-status
grpcio-status==1.78.0
    # via google-api-core
gunicorn==25.1.0
    # via
    #   functions-framework
    #   uvicorn-worker
h11==0.16.0
    # via uvicorn
idna==3.11
    # via
    #   anyio
    #   requests
itsdangerous==2.2.0
    # via flask
jinja2==3.1.6
    # via flask
markupsafe==3.0.3
    # via
    #   flask
    #   jinja2
    #   werkzeug
packaging==26.0
    # via
    #   deprecation
    #   gunicorn
proto-plus==1.27.1
    # via
    #   google-api-core
    #   google-cloud-documentai
protobuf==6.33.5
    # via
    #   google-api-core
    #   google-cloud-documentai
    #   googleapis-common-protos
    #   grpcio-status
    #   proto-plus
pyasn1==0.6.2
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pycparser==3.0
    # via cffi
pydantic==2.12.5
    # via
    #   -r functions/ocr_trigger/requirements.in
    #   pydantic-settings
pydantic-core==2.41.5
    # via pydantic
pydantic-settings==2.13.0
    # via -r functions/ocr_trigger/requirements.in
python-dotenv==1.2.1
    # via pydantic-settings
requests==2.32.5
    # via google-api-core
rsa==4.9.1
    # via google-auth
starlette==0.52.1
    # via functions-framework
typing-extensions==4.15.0
    # via
    #   anyio
    #   cryptography
    #   exceptiongroup
    #   grpcio
    #   pydantic
    #   pydantic-core
    #   starlette
    #   typing-inspection
    #   uvicorn
typing-inspection==0.4.2
    # via
    #   pydantic
    #   pydantic-settings
urllib3==2.6.3
    # via requests
uvicorn==0.41.0
    # via
    #   functions-framework
    #   uvicorn-worker
uvicorn-worker==0.4.0
    # via functions-framework
watchdog==6.0.0
    # via functions-framework
werkzeug==3.1.5
    # via
    #   flask
    #   functions-framework